requirements.txt
# Core dependencies
python-telegram-bot==13.7
pandas>=1.5.0
numpy<2.0.0,>=1.24.0
pytz>=2021.3
requests>=2.27.1
matplotlib>=3.5.1
python-dotenv>=0.19.2
ccxt>=1.77.55
ta>=0.10.1
MetaTrader5>=5.0.44

# AI-related dependencies
tensorflow>=2.10.0
scikit-learn>=1.0.2
keras>=2.10.0
joblib>=1.1.0
openai>=1.0.0
